@@ -1,60 +1,60 @@
-const Collection = require('../util/Collection');
+import Collection from '../util/Collection.js'
 
 class Channel {
+  /**
+   * A Twitch channel.
+   * @param {EmoteFetcher} fetcher - The emote fetcher.
+   * @param {number} id - ID of the channel.
+   */
+  constructor (fetcher, id) {
     /**
-     * A Twitch channel.
-     * @param {EmoteFetcher} fetcher - The emote fetcher.
-     * @param {int} id - ID of the channel.
+     * The emote fetcher.
+     * @type {EmoteFetcher}
      */
-    constructor(fetcher, id) {
-        /**
-         * The emote fetcher.
-         * @type {EmoteFetcher}
-         */
-        this.fetcher = fetcher;
-
-        /**
-         * The ID of this channel.
-         * For the global channel, the ID will be null.
-         * @type {?int}
-         */
-        this.channel_id = id;
-
-        /**
-         * Cached emotes belonging to this channel.
-         * @type {Collection<string, Emote>}
-         */
-        this.emotes = new Collection();
-    }
-
-    /* There are tests that those are returning Collections, but c8 doesn't get it. */
-    /* c8 ignore start */
+    this.fetcher = fetcher
 
     /**
-     * Fetches the BTTV emotes for this channel.
-     * @returns {Promise<Collection<string, BTTVEmote>>}
+     * The ID of this channel.
+     * For the global channel, the ID will be null.
+     * @type {?number}
      */
-    fetchBTTVEmotes() {
-        return this.fetcher.fetchBTTVEmotes(this.id);
-    }
+    this.channel_id = id || null
 
     /**
-     * Fetches the FFZ emotes for this channel.
-     * @returns {Promise<Collection<string, FFZEmote>>}
+     * Cached emotes belonging to this channel.
+     * @type {Collection<string, Emote>}
      */
-    fetchFFZEmotes() {
-        return this.fetcher.fetchFFZEmotes(this.id);
-    }
+    this.emotes = new Collection()
+  }
 
-    /**
-     * Fetches the 7TV emotes for this channel.
-     * @returns {Promise<Collection<string, SevenTVEmote>>}
-     */
-    fetchSevenTVEmotes() {
-        return this.fetcher.fetchSevenTVEmotes(this.id);
-    }
+  /* There are tests that those are returning Collections, but c8 doesn't get it. */
+  /* c8 ignore start */
 
-    /* c8 ignore end */
+  /**
+   * Fetches the BTTV emotes for this channel.
+   * @returns {Promise<Collection<string, BTTVEmote>>} - A promise that resolves to a collection of BTTVEmotes.
+   */
+  fetchBTTVEmotes () {
+    return this.fetcher.fetchBTTVEmotes(this.id)
+  }
+
+  /**
+   * Fetches the FFZ emotes for this channel.
+   * @returns {Promise<Collection<string, FFZEmote>>} - A promise that resolves to a collection of FFZEmotes.
+   */
+  fetchFFZEmotes () {
+    return this.fetcher.fetchFFZEmotes(this.id)
+  }
+
+  /**
+   * Fetches the 7TV emotes for this channel.
+   * @returns {Promise<Collection<string, SevenTVEmote>>} - A promise that resolves to a collection of SevenTVEmotes.
+   */
+  fetchSevenTVEmotes () {
+    return this.fetcher.fetchSevenTVEmotes(this.id)
+  }
+
+  /* c8 ignore end */
 }
 
-module.exports = Channel;
+export default Channel
