@@ -1,13 +1,13 @@
-const BTTVEmote = require('./BTTVEmote');
-const FFZEmote = require('./FFZEmote');
-const SevenTVEmote = require('./SevenTVEmote');
-const TwitchEmote = require('./TwitchEmote');
+import BTTVEmote from './BTTVEmote.js'
+import FFZEmote from './FFZEmote.js'
+import SevenTVEmote from './SevenTVEmote.js'
+import TwitchEmote from './TwitchEmote.js'
 
 class EmoteTypeMapper {
-    static getClassByType(type) {
-        const classMap = { BTTVEmote, FFZEmote, SevenTVEmote, TwitchEmote };
-        return classMap[type] || null;
-    }
+  static getClassByType (type) {
+    const classMap = { BTTVEmote, FFZEmote, SevenTVEmote, TwitchEmote }
+    return classMap[type] || null
+  }
 }
 
-module.exports = EmoteTypeMapper;
+export default EmoteTypeMapper
