@@ -1,12 +1,38 @@
-module.exports = {
-    BTTVEmote: require('./struct/BTTVEmote'),
-    Channel: require('./struct/Channel'),
-    Collection: require('./util/Collection'),
-    Constants: require('./util/Constants'),
-    Emote: require('./struct/Emote'),
-    EmoteFetcher: require('./struct/EmoteFetcher'),
-    EmoteParser: require('./struct/EmoteParser'),
-    FFZEmote: require('./struct/FFZEmote'),
-    SevenTVEmote: require('./struct/SevenTVEmote'),
-    TwitchEmote: require('./struct/TwitchEmote')
-};
+/// <reference types="../typings/index.d.ts" />
+import BTTVEmote from './struct/BTTVEmote.js'
+import Channel from './struct/Channel.js'
+import Collection from './util/Collection.js'
+import Constants from './util/Constants.js'
+import Emote from './struct/Emote.js'
+import EmoteFetcher from './struct/EmoteFetcher.js'
+import EmoteParser from './struct/EmoteParser.js'
+import FFZEmote from './struct/FFZEmote.js'
+import SevenTVEmote from './struct/SevenTVEmote.js'
+import TwitchEmote from './struct/TwitchEmote.js'
+
+export {
+  BTTVEmote,
+  Channel,
+  Collection,
+  Constants,
+  Emote,
+  EmoteFetcher,
+  EmoteParser,
+  FFZEmote,
+  SevenTVEmote,
+  TwitchEmote,
+}
+
+// Default export for compatibility
+export default {
+  BTTVEmote,
+  Channel,
+  Collection,
+  Constants,
+  Emote,
+  EmoteFetcher,
+  EmoteParser,
+  FFZEmote,
+  SevenTVEmote,
+  TwitchEmote,
+}
